requirements-dev.txt
-r requirements.txt
django-debug-toolbar==1.11
django-filter==2.1.0
djangorestframework==3.9.1
flake8==3.6.0
idna==2.8
idna-ssl==1.1.0
lxml==4.3.0
mccabe==0.6.1
multidict==4.5.2
pep8==1.7.1
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.9
requests==2.21.0
sqlparse==0.2.4
typing-extensions==3.7.2
urllib3==1.24.1
yarl==1.3.0
